task: added notification and email config

+++ app/Http/Controllers/ExpenseController.php
@@ -2,56 +2,67 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Expense;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
+use App\Mail\ExpenseCreatedMail;
+use App\Mail\ExpenseUpdatedMail;
+use App\Mail\ExpenseApprovedMail;
+use App\Mail\ExpenseRejectedMail;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\StoreExpenseRequest;
 use App\Http\Requests\UpdateExpenseRequest;
+use Illuminate\Support\Facades\Notification;
+use App\Notifications\ExpenseCreatedNotification;
+use App\Notifications\ExpenseDeletedNotification;
+use App\Notifications\ExpenseUpdatedNotification;
+use App\Notifications\ExpenseApprovedNotification;
+use App\Notifications\ExpenseRejectedNotification;
 
 class ExpenseController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+    
     public function index(Request $request)
     {
-        //
         $userId = auth()->user()->id;
         $dateFilter = $request->date_filter;
-        $query = Expense::query()->where('user_id', $userId);
-        
-        switch($dateFilter){
-            
-            case 'today':
-                $query->whereDate('created_at',Carbon::today());
-                break;
-            case 'yesterday':
-                $query->wheredate('created_at',Carbon::yesterday());
-                break;
-            case 'this_week':
-                $query->whereBetween('created_at',[Carbon::now()->startOfWeek(),Carbon::now()->endOfWeek()]);
-                break;
-            case 'last_week':
-                $query->whereBetween('created_at',[Carbon::now()->subWeek(),Carbon::now()]);
-                break;
-            case 'this_month':
-                $query->whereMonth('created_at',Carbon::now()->month);
-                break;
-            case 'last_month':
-                $query->whereMonth('created_at',Carbon::now()->subMonth()->month);
-                break;
-            case 'this_year':
-                $query->whereYear('created_at',Carbon::now()->year);
-                break;
-            case 'last_year':
-                $query->whereYear('created_at',Carbon::now()->subYear()->year);
-                break;                       
-            }
-        $expenses = $query->paginate(5);
-        return view('expenses.index', compact('expenses', 'dateFilter'));
+        $statusFilter = $request->status_filter;
+        $search = $request->search;
+
+        if(auth()->user()->hasRole('Administrator')){
+            $query = Expense::query();
+        }
+        else{
+            $query = Expense::query()->where('user_id', $userId);
+        }
+        $query = match($dateFilter) {
+            'today' => $query->whereDate('created_at', Carbon::today()),
+            'yesterday' => $query->whereDate('created_at', Carbon::yesterday()),
+            'this_week' => $query->whereBetween('created_at', [Carbon::now()->startOfWeek(), Carbon::now()->endOfWeek()]),
+            'last_week' => $query->whereBetween('created_at', [Carbon::now()->subWeek()->startOfWeek(), Carbon::now()->subWeek()->endOfWeek()]),
+            'this_month' => $query->whereMonth('created_at', Carbon::now()->month),
+            'last_month' => $query->whereMonth('created_at', Carbon::now()->subMonth()->month),
+            'this_year' => $query->whereYear('created_at', Carbon::now()->year),
+            'last_year' => $query->whereYear('created_at', Carbon::now()->subYear()->year),
+            default => $query,
+        };
+        if ($statusFilter) {
+            $query->where('status', $statusFilter);
+        }
+        if ($search) {
+            $query->where('title', 'LIKE', '%' . $search . '%');
+        }
+
+        $expenses = $query->latest()->paginate(5)->appends(['date_filter' => $dateFilter, 'status_filter' => $statusFilter, 'search'=> $search]);
+        return view('expenses.index', compact('expenses', 'dateFilter', 'statusFilter', 'search'));
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -68,23 +79,32 @@ class ExpenseController extends Controller
     public function store(StoreExpenseRequest $request)
     {
         //
-        $expense = Expense::create([
-            'title'       => $request->title,
-            'description' => $request->description,
-            'quantity' => $request->quantity,
-            'unit_price' => $request->unit_price,
-            'category' => $request->category,
-            'user_id'     => auth()->id(),
-        ]);
+        $user = auth()->user();
+        $validated = $request->validated();
         if($request->file('receipt')){
-            $extension      = $request->file('receipt')->extension();
+            $extension = $request->file('receipt')->extension();
             $contents = file_get_contents($request->file('receipt'));
             $filename = Str::random(25);
-            $path     = "attachments/$filename.$extension";
+            $path = "attachments/$filename.$extension";
             Storage::disk('public')->put($path, $contents);
-            $expense->update(['receipt' => $path]);
+            $validated['receipt'] = $path;
         }
-        return redirect()->route('expense.index');
+        // create a new expense record
+        $expense = auth()->user()->expenses()->create( $validated );   
+
+        // Notify the admins
+        // $user->notify(new ExpenseCreatedNotification($expense));
+        $admins = User::whereHas('roles', function ($query) {$query->where('name', 'Administrator');})->get();
+  
+        Notification::send($admins, new ExpenseCreatedNotification($expense));
+
+        // Send mail admins
+        // Mail::to($request->user())->send(new ExpenseCreated($expense));
+        foreach ($admins as $recipient) {
+            Mail::to($recipient)->send(new ExpenseCreatedMail($expense));
+        }
+
+        return redirect()->route('expense.index')->with('success','Expense created successfully');
 
     }
 
@@ -112,19 +132,71 @@ class ExpenseController extends Controller
     public function update(UpdateExpenseRequest $request, Expense $expense)
     {
         //
-        $expense->update($request->except('receipt'));
+        //$expense->update($request->validated());
+        //$validatedData = $request->validated();
 
-        if ($request->file('receipt')) {
-            Storage::disk('public')->delete($expense->receipt);
-            $extension      = $request->file('receipt')->extension();
-            $contents = file_get_contents($request->file('receipt'));
-            $filename = Str::random(25);
-            $path     = "attachments/$filename.$extension";
-            Storage::disk('public')->put($path, $contents);
-            $expense->update(['receipt' => $path]);
+        if (auth()->user()->hasRole('Administrator')) {
+            
+            $request->validate([
+                'status' => 'required|in:pending,approved,rejected',
+            ]);
+            $expense->update(['status' => $request->status]);
+            if ($expense['status'] === 'approved') {
+                $expense->user->notify(new ExpenseApprovedNotification($expense));
+                Mail::to($request->user())->send(new ExpenseApprovedMail($expense));
+            } elseif ($expense['status'] === 'rejected') {
+                $expense->user->notify(new ExpenseRejectedNotification($expense));
+                Mail::to($request->user())->send(new ExpenseRejectedMail($expense));
+            }
+        }
+        else{
+            $expense->update($request->except('receipt'));
+            if($request->file('receipt')) {
+                Storage::disk('public')->delete($expense->receipt);
+                $extension      = $request->file('receipt')->extension();
+                $contents = file_get_contents($request->file('receipt'));
+                $filename = Str::random(25);
+                $path     = "attachments/$filename.$extension";
+                Storage::disk('public')->put($path, $contents);
+                $expense->update(['receipt' => $path]);
+            }
            
         }
-        return redirect(route('expense.index'));
+        // if ($request->hasFile('receipt')) {
+        //     // Delete the previous receipt file if it exists
+        //     if ($expense->receipt && Storage::disk('public')->exists($expense->receipt)) {
+        //         Storage::disk('public')->delete($expense->receipt);
+        //     }
+    
+        //     // Store the new receipt file
+        //     $extension = $request->file('receipt')->extension();
+        //     $contents = file_get_contents($request->file('receipt'));
+        //     $filename = Str::random(25);
+        //     $path = "attachments/$filename.$extension";
+        //     Storage::disk('public')->put($path, $contents);
+    
+        //     // Update the 'receipt' field with the new file path
+        //     $validatedData['receipt'] = $path;
+        // }
+    
+        // $expense->update($validatedData);
+        // $expense->status = $request->input('status'); // Update the status field
+        // $expense->save();
+        //auth()->user()->notify(new ExpenseUpdatedNotification($expense));
+        // Send notification
+        //$admins = User::whereHas('roles', function ($query) {$query->where('name', 'admin');})->get();
+  
+        //Notification::send($admins, new ExpenseUpdatedNotification($expense));
+
+        //Send mail
+
+        //foreach ($admins as $recipient) {
+          //  Mail::to($recipient)->send(new ExpenseUpdatedMail($expense));
+        //}
+
+        
+        return redirect()->route('expense.index')->with('success','Expense Updated successfully');
+
     }
 
     /**
@@ -134,6 +206,26 @@ class ExpenseController extends Controller
     {
         //
         $expense->delete();
+        //auth()->user()->notify(new ExpenseDeletedNotification($expense));
+
         return redirect(route("expense.index"));
     }
+
+    // public function updateStatus(Expense $expense, $status)
+    // {
+    //     // if (!Auth::user()->hasRole('Administrator')) {
+    //     //     abort(403, 'Unauthorized action.');
+    //     // }
+
+    //     $validStatus = ['approved', 'rejected'];
+    //     if (!in_array($status, $validStatus)) {
+    //         abort(400, 'Invalid status value.');
+    //     }
+
+    //     $expense->update(['status' => $status]);
+
+    //     // Notify the user or perform any other necessary actions
+
+    //     return redirect()->route('expense.show', $expense->id);
+    // }
 }
